Adapt for the package.js which is shipped by gjs
package.js requires the launched script to be named after the
app it launches (otherwise pkg.name reports wrong values).
Also, the sources need to be compiled in GResource now,
and imports are no longer automagic.

This does not affect the command line (because a symlink is
installed in $(bindir) as "gnome-sound-recorder") and does
not affect launching from srcdir (the real JS files are loaded)

https://bugzilla.gnome.org/show_bug.cgi?id=732847